Code generation project, and first-use-page with create page admin event handling.

+++ ProFind/Lib/AdminNS/Views/CRUDPages/ProfessionNS/ListPage/List-Page.xaml.cs
@@ -19,7 +19,7 @@ namespace ProFind.Lib.AdminNS.Views.CRUDPages.ProfessionNS.ListPage
         }
         private async void InitializeData()
         {
-            int idNo = Id1.IdPfs;
+            int idNo = (int)Id1.IdPfs;
             await APIConnection.GetConnection.GetProfessionAsync(idNo);
         }
     }
